added sling tv designation to the cable channels that are on sling I am coding like a boss. the boss of a half a bottle of wine.
There was a requirement to have sling tv on shows that are shown on
cable regularly. I added a method to the search.js to add sling shows if
they match a list. there is a more elegant way to do this but for MVP
this is a good piece of work that can be debugged later cleanly.