Upgraded django to 1.9 and python to 3.5.1
In order to get a JSON data type in postgress compatible with the django
ORM an upgrade was necessary. The application is currently working and
there are some problems with the angular. But we should be good to go
with a target finish of backend changes by the end of this week.